+++ Application/Services/StudyGroupServices.cs
@@ -10,28 +10,33 @@ public class StudyGroupServices(IStudyGroupRepository studyGroupRepository, IUni
 {
     public async Task<List<StudyGroupDto>> FetchStudyGroupsFromBackendAsync()
     {
-        var studyGroupList = await studyGroupRepository.GetStudyGroupListAsync();
+        var studyGroupList = await studyGroupRepository.GetStudyGroupListAsync(1);
         return studyGroupList.ToStudyGroupsDto();
     }
 
     public async Task AddStudyGroup(StudyGroupDto studyGroupDto)
     {
-        var studyGroup = StudyGroup.CreateStudyGroup(studyGroupDto.Name);
-        await studyGroupRepository.AddAsync(studyGroup);
+        var studyGroup = StudyGroup.CreateStudyGroup(studyGroupDto.Id, studyGroupDto.Name);
+        await studyGroupRepository.AddAsync(studyGroup, 1);
         await unitOfWork.SaveChangesAsync();
     }
 
-    public async Task EditStudyGroup(StudyGroupDto oldStudyGroupDto, StudyGroupDto newStudyGroupDto)
+    public async Task EditStudyGroup(StudyGroupDto studyGroupDto)
     {
-        var oldStudyGroup = StudyGroup.CreateStudyGroup(oldStudyGroupDto.Name);
-        var newStudyGroup = StudyGroup.CreateStudyGroup(newStudyGroupDto.Name);
-        await studyGroupRepository.UpdateAsync(oldStudyGroup, newStudyGroup);
+        var studyGroup = await studyGroupRepository.GetStudyGroupByIdAsync(studyGroupDto.Id);
+        if (studyGroup is null)
+            throw new  ArgumentException($"StudyGroup with id {studyGroupDto.Id} not found");
+        if (studyGroupDto.Name != studyGroup.Name)
+            studyGroup.SetName(studyGroupDto.Name);
+        await studyGroupRepository.UpdateAsync(studyGroup);
         await unitOfWork.SaveChangesAsync();
     }
 
     public async Task DeleteStudyGroup(StudyGroupDto studyGroupDto)
     {
-        var studyGroup = StudyGroup.CreateStudyGroup(studyGroupDto.Name);
+        var studyGroup = await studyGroupRepository.GetStudyGroupByIdAsync(studyGroupDto.Id);
+        if (studyGroup is null)
+            throw new  ArgumentException($"StudyGroup with id {studyGroupDto.Id} not found");
         await studyGroupRepository.Delete(studyGroup);
         await unitOfWork.SaveChangesAsync();
     }
